[LOGIN] 전역함수 추가
